+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # VocalFusion AVS Setup Change Log
 
+## 3.0.0
+
+  * Added support for xvf3610-ua and xvf3615-ua
+  * Renamed xvf3615 device as xvf3615-int
+  * Renamed xvf3610 device as xvf3610-int
+
 ## 2.3.0
 
   * Added support for xvf3615